structure: moved directories and standardised filenames
