+++ packages/subgraph/manifest/subgraph.placeholder.yaml
@@ -50,7 +50,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.5
       language: wasm/assemblyscript
-      file: ./src/plugin/token-voting.ts
+      file: ./src/plugin/plugin.ts
       entities:
         - TokenVotingPlugin
         - TokenVotingMember
