+++ packages/subgraph/src/osx/pluginSetupProcessor.ts
@@ -3,7 +3,7 @@ import {TokenVotingPlugin as TokenVotingPluginEntity} from '../../generated/sche
 import {TokenVoting} from '../../generated/templates';
 import {TokenVoting as TokenVotingPluginContract} from '../../generated/templates/TokenVoting/TokenVoting';
 import {PLUGIN_REPO_ADDRESS} from '../../imported/repo-address';
-import {identifyAndFetchOrCreateERC20TokenEntity} from '../utils/tokens/erc20';
+import {identifyAndFetchOrCreateERC20TokenEntity} from '../utils/erc20';
 import {generatePluginEntityId} from '@aragon/osx-commons-subgraph';
 import {
   Address,
