--- packages/subgraph/tests/utils/events/index.ts
@@ -1,1 +1,0 @@
-export * from './osx';
