+++ packages/subgraph/tests/utils/osx.ts
@@ -1,7 +1,7 @@
 import {
   InstallationPrepared,
   InstallationPreparedPreparedSetupDataStruct,
-} from '../../../generated/PluginSetupProcessor/PluginSetupProcessor';
+} from '../../generated/PluginSetupProcessor/PluginSetupProcessor';
 import {Address, Bytes, ethereum} from '@graphprotocol/graph-ts';
 import {newMockEvent} from 'matchstick-as';
